+++ packages/firebase-hooks/docs/stories/index.story.js
@@ -17,3 +17,4 @@ export * from './useDocOnce.story';
 export * from './useHttpAuthMutation.story';
 export * from './useHttpAuthQuery.story';
 export * from './useUserClaims.story';
+export * from './useCallableQuery.story';