+++ packages/forms/src/components/CountryPicker/index.d.ts
@@ -32,6 +32,10 @@ export interface CountryPickerProps {
     name?: string;
     /** Specifies that the picker must be filled out before submitting the form. */
     error?: ReactNode;
+    /** A component that is displayed when no search matches are found */
+    emptyComponent?: ReactNode;
+    /** Specifies the ISO codes of the countries available for selection. */
+    countries?: string[];
 }
 
 export default function (props: CountryPickerProps): JSX.Element | null;