+++ packages/listview/src/components/AuditLogs/header.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable react/jsx-one-expression-per-line */
 import React, { useContext, useRef, useState } from 'react';
 import { Download } from '@rainbow-modules/icons';
 import { DatePickerModal, Option } from 'react-rainbow-components';
@@ -18,17 +17,20 @@ import context from './context';
 import SeveritySelect from './severitySelect';
 import formatDates from './helpers/formatDates';
 import FilterButtonLabel from './filterButtonLabel';
+import DownloadModal from '../DownloadModal';
 
 const ToolBar = () => {
     const [isFilterOpen, setIsFilterOpen] = useState<boolean>(false);
     const [isDatePickerOpen, setIsDatePickerOpen] = useState<boolean>(false);
+    const [isDownloadOpen, setIsDownloadOpen] = useState<boolean>(false);
     // TODO: fix this state types
     const [dateValue, setDateValue] = useState<any>();
-    const { filters, updateFilters } = useContext(context);
+    const { filters, updateFilters, getDownloadData } = useContext(context);
     const { severity, dateRange } = filters;
 
     const filterButtonRef = useRef<ButtonHandle>();
-    const dropdownRef = useRef<HTMLDivElement>(null);
+
+    const closeLabelFilters = () => setIsFilterOpen(false);
 
     const toggleLabelFilters = () => setIsFilterOpen(!isFilterOpen);
 
@@ -60,8 +62,13 @@ const ToolBar = () => {
     return (
         <>
             <SeveritySelect severity={severity} handleFilterChange={handleFilterChange} />
-            <StyledFilterPicklist name="dateRange" onChange={handleDateChange} value={dateRange}>
-                <Option name="all" label="All Dates" />
+            <StyledFilterPicklist
+                name="dateRange"
+                onChange={handleDateChange}
+                onClick={closeLabelFilters}
+                value={dateRange}
+            >
+                <Option name="all" label="All Time" />
                 <Option name="today" label="Today" />
                 <Option name="this-week" label="This Week" />
                 <Option name="this-month" label="This Month" />
@@ -73,12 +80,16 @@ const ToolBar = () => {
                 onClick={toggleLabelFilters}
                 ref={filterButtonRef}
             />
-            <StyledFilterButtonIcon icon={<Download />} variant="neutral" />
+            <StyledFilterButtonIcon
+                icon={<Download />}
+                variant="neutral"
+                onClick={() => setIsDownloadOpen(true)}
+            />
             <InternalOverlay
                 isVisible={isFilterOpen}
                 triggerElementRef={filterButtonRef.current?.htmlElementRef}
             >
-                <FilterDropdown ref={dropdownRef} />
+                <FilterDropdown close={closeLabelFilters} />
             </InternalOverlay>
             <DatePickerModal
                 isOpen={isDatePickerOpen}
@@ -87,6 +98,13 @@ const ToolBar = () => {
                 value={dateValue}
                 onChange={handleDatePickerChange}
                 onRequestClose={() => setIsDatePickerOpen(false)}
+            />
+            <DownloadModal
+                id="default-download-modal"
+                isOpen={isDownloadOpen}
+                onRequestClose={() => setIsDownloadOpen(false)}
+                maxEntries={10}
+                onDownload={getDownloadData}
             />
         </>
     );