+++ packages/listview/src/components/AuditLogs/helpers/index.ts
@@ -2,3 +2,5 @@ export { default as getCurrentDay } from './getCurrentDay';
 export { default as getCurrentWeek } from './getCurrentWeek';
 export { default as getCurrentMonth } from './getCurrentMonth';
 export { default as getCustomDates } from './getCustomDates';
+export { default as getFormattedCode } from './getFormattedCode';
+export { default as prepareForDownload } from './prepareForDownload';