+++ packages/listview/src/components/AuditLogs/types.ts
@@ -1,10 +1,23 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { ReactNode, RefObject } from 'react';
+import { ReactNode, RefObject, CSSProperties } from 'react';
+import { Max } from '../DownloadModal';
+
+export type Severity = 'debug' | 'info' | 'warning' | 'error';
+
+export type LabelFilter = {
+    [name: string]: string | undefined;
+};
+
+export interface Filters {
+    severity?: Severity[];
+    dateRange?: DateRange;
+    labels?: LabelFilter;
+}
 
 export interface AuditLogsProps {
     collectionPath: string;
     labels?: string[];
-    defaultFilter?: Record<string, any>;
+    defaultFilter?: Filters;
 }
 
 export interface SeverityProps {
@@ -22,15 +35,7 @@ export type ContextType = {
     filters: Filters;
     labels: string[];
     updateFilters: (newFilters: Filters) => void;
-};
-
-export type Severity = 'debug' | 'info' | 'warning' | 'error';
-
-export type LabelFilter = {
-    name: string;
-    value: string;
-    index: number;
-    label?: string;
+    getDownloadData?: (max: Max, format: string) => Promise<Array<Record<string, unknown>>>;
 };
 
 export interface SeveritySelectProps {
@@ -43,15 +48,28 @@ export interface DateRange {
     label?: string;
 }
 
-export interface Filters {
-    severity?: Severity[];
-    dateRange?: DateRange;
-    labels?: Record<string, string[]>;
-}
-
 export interface PicklistValue {
     label?: string;
     name?: string | number;
     icon?: ReactNode;
     value?: any;
+}
+
+export interface ClientFilterTableProps {
+    filters: Filters;
+    collection: string;
+    query?: (ref: Record<string, unknown>) => void;
+}
+
+export interface FilterDropdownProps {
+    close: () => void;
+}
+
+export interface BaseProps {
+    className?: string;
+    style?: CSSProperties;
+}
+
+export interface IconProps extends BaseProps {
+    title?: string;
 }