+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.49.0](https://github.com/nexxtway/rainbow-modules/compare/v0.48.0...v0.49.0) (2022-03-06)
+
+### Bug Fixes
+
+-   improved country picker ([#542](https://github.com/nexxtway/rainbow-modules/issues/542)) ([a326a89](https://github.com/nexxtway/rainbow-modules/commit/a326a895c47236dfa688eea054fe3a9d82debfac))
+
 # [0.48.0](https://github.com/nexxtway/rainbow-modules/compare/v0.47.3...v0.48.0) (2022-01-22)
 
 **Note:** Version bump only for package @rainbow-modules/forms