+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.49.0](https://github.com/nexxtway/rainbow-modules/compare/v0.48.0...v0.49.0) (2022-03-06)
+
+### Features
+
+-   added new icons ([#520](https://github.com/nexxtway/rainbow-modules/issues/520)) ([2a45e61](https://github.com/nexxtway/rainbow-modules/commit/2a45e61222fd27747f1096d2af5ff99254bdac89))
+-   implement AuditLogs component ([#521](https://github.com/nexxtway/rainbow-modules/issues/521)) ([15dc915](https://github.com/nexxtway/rainbow-modules/commit/15dc9158a0ec64dc5c212e5c5826d6c95f2cf658))
+
 # [0.48.0](https://github.com/nexxtway/rainbow-modules/compare/v0.47.3...v0.48.0) (2022-01-22)
 
 **Note:** Version bump only for package @rainbow-modules/icons